extract plot package
