+++ src/main/python/plotlyst/view/widget/plot/editor.py
@@ -1,0 +1,641 @@
+"""
+Plotlyst
+Copyright (C) 2021-2024  Zsolt Kovari
+
+This file is part of Plotlyst.
+
+Plotlyst is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Plotlyst is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+"""
+
+from functools import partial
+from typing import Set, Dict, Optional
+
+import qtanim
+from PyQt6.QtCore import pyqtSignal, Qt, QSize, QTimer
+from PyQt6.QtGui import QColor, QCursor
+from PyQt6.QtWidgets import QWidget, QFrame
+from overrides import overrides
+from qthandy import bold, flow, incr_font, \
+    margins, ask_confirmation, italic, retain_when_hidden, transparent, \
+    clear_layout, vspacer, decr_font, decr_icon, hbox, spacer, sp, pointy, incr_icon, translucent
+from qthandy.filter import VisibilityToggleEventFilter, OpacityEventFilter
+from qtmenu import MenuWidget
+
+from src.main.python.plotlyst.common import RELAXED_WHITE_COLOR
+from src.main.python.plotlyst.core.domain import Novel, Plot, PlotValue, PlotType, Character, PlotPrinciple, \
+    PlotPrincipleType, PlotProgressionItem, \
+    PlotProgressionItemType
+from src.main.python.plotlyst.env import app_env
+from src.main.python.plotlyst.event.core import EventListener, Event, emit_event
+from src.main.python.plotlyst.event.handler import event_dispatchers
+from src.main.python.plotlyst.events import CharacterChangedEvent, CharacterDeletedEvent, StorylineCreatedEvent, \
+    StorylineRemovedEvent, StorylineCharacterAssociationChanged
+from src.main.python.plotlyst.service.persistence import RepositoryPersistenceManager, delete_plot
+from src.main.python.plotlyst.settings import STORY_LINE_COLOR_CODES
+from src.main.python.plotlyst.view.common import action, fade_out_and_gc, ButtonPressResizeEventFilter, \
+    insert_before_the_end, label
+from src.main.python.plotlyst.view.dialog.novel import PlotValueEditorDialog
+from src.main.python.plotlyst.view.dialog.utility import IconSelectorDialog
+from src.main.python.plotlyst.view.generated.plot_editor_widget_ui import Ui_PlotEditor
+from src.main.python.plotlyst.view.generated.plot_widget_ui import Ui_PlotWidget
+from src.main.python.plotlyst.view.icons import IconRegistry, avatars
+from src.main.python.plotlyst.view.widget.button import SecondaryActionPushButton
+from src.main.python.plotlyst.view.widget.characters import CharacterAvatar, CharacterSelectorMenu
+from src.main.python.plotlyst.view.widget.labels import PlotValueLabel
+from src.main.python.plotlyst.view.widget.plot.matrix import StorylinesImpactMatrix
+from src.main.python.plotlyst.view.widget.plot.principle import PlotPrincipleSelectorMenu, PlotPrincipleEditor, \
+    PrincipleSelectorObject, GenrePrincipleSelectorDialog
+from src.main.python.plotlyst.view.widget.plot.progression import PlotEventsTimeline
+from src.main.python.plotlyst.view.widget.tree import TreeView, ContainerNode
+from src.main.python.plotlyst.view.widget.utility import ColorPicker
+
+
+class PlotNode(ContainerNode):
+    def __init__(self, plot: Plot, parent=None):
+        super(PlotNode, self).__init__(plot.text, parent)
+        self._plot = plot
+        self.setPlusButtonEnabled(False)
+        incr_font(self._lblTitle)
+        margins(self._wdgTitle, top=5, bottom=5)
+
+        self.refresh()
+
+    def plot(self) -> Plot:
+        return self._plot
+
+    def refresh(self):
+        if self._plot.icon:
+            self._icon.setIcon(IconRegistry.from_name(self._plot.icon, self._plot.icon_color))
+            self._icon.setVisible(True)
+        else:
+            self._icon.setHidden(True)
+
+        self._lblTitle.setText(self._plot.text)
+
+
+class PlotTreeView(TreeView, EventListener):
+    plotSelected = pyqtSignal(Plot)
+    plotRemoved = pyqtSignal(Plot)
+
+    def __init__(self, novel: Novel, parent=None):
+        super().__init__(parent)
+        self._novel = novel
+        self._plots: Dict[Plot, PlotNode] = {}
+        self._characterNodes: Dict[Character, ContainerNode] = {}
+        self._selectedPlots: Set[Plot] = set()
+
+        self.refresh()
+
+        event_dispatchers.instance(self._novel).register(self, StorylineCharacterAssociationChanged)
+
+    @overrides
+    def event_received(self, event: Event):
+        self.refresh()
+
+    def refresh(self):
+        self._selectedPlots.clear()
+        self._characterNodes.clear()
+        self._plots.clear()
+        clear_layout(self._centralWidget)
+
+        characters = [x.character(self._novel) for x in self._novel.plots if x.character_id]
+        characters_set = set(characters)
+        characters_set.discard(None)
+        if len(characters_set) > 1:
+            for character in characters:
+                if character in self._characterNodes.keys():
+                    continue
+                self._characterNodes[character] = ContainerNode(character.name, avatars.avatar(character),
+                                                                readOnly=True)
+                self._centralWidget.layout().addWidget(self._characterNodes[character])
+
+        for plot in self._novel.plots:
+            wdg = self.__initPlotWidget(plot)
+            if plot.character_id and self._characterNodes:
+                character = plot.character(self._novel)
+                self._characterNodes[character].addChild(wdg)
+            else:
+                self._centralWidget.layout().addWidget(wdg)
+
+        self._centralWidget.layout().addWidget(vspacer())
+
+    def refreshPlot(self, plot: Plot):
+        self._plots[plot].refresh()
+
+    def refreshCharacters(self):
+        self.refresh()
+
+    def addPlot(self, plot: Plot):
+        wdg = self.__initPlotWidget(plot)
+        self._centralWidget.layout().insertWidget(self._centralWidget.layout().count() - 1, wdg)
+
+    def selectPlot(self, plot: Plot):
+        self._plots[plot].select()
+        wdg = self._plots[plot]
+        self._plotSelectionChanged(wdg, wdg.isSelected())
+
+    def removePlot(self, plot: Plot):
+        self._removePlot(self._plots[plot])
+
+    def clearSelection(self):
+        for plot in self._selectedPlots:
+            self._plots[plot].deselect()
+        self._selectedPlots.clear()
+
+    def _plotSelectionChanged(self, wdg: PlotNode, selected: bool):
+        if selected:
+            self.clearSelection()
+            self._selectedPlots.add(wdg.plot())
+            QTimer.singleShot(10, lambda: self.plotSelected.emit(wdg.plot()))
+        elif wdg.plot() in self._selectedPlots:
+            self._selectedPlots.remove(wdg.plot())
+
+    def _removePlot(self, wdg: PlotNode):
+        plot = wdg.plot()
+        if not ask_confirmation(f"Delete plot '{plot.text}'?", self._centralWidget):
+            return
+        if plot in self._selectedPlots:
+            self._selectedPlots.remove(plot)
+        self._plots.pop(plot)
+
+        characterNode = None
+        if plot.character_id and self._characterNodes:
+            character = plot.character(self._novel)
+            characterNode = self._characterNodes[character]
+            if len(characterNode.childrenWidgets()) == 1:
+                self._characterNodes.pop(character)  # remove parent too
+            else:
+                characterNode = None  # keep parent
+
+        fade_out_and_gc(wdg.parent(), wdg)
+        if characterNode:
+            fade_out_and_gc(self._centralWidget, characterNode)
+
+        self.plotRemoved.emit(wdg.plot())
+
+    def __initPlotWidget(self, plot: Plot) -> PlotNode:
+        if plot not in self._plots.keys():
+            wdg = PlotNode(plot)
+            wdg.selectionChanged.connect(partial(self._plotSelectionChanged, wdg))
+            wdg.deleted.connect(partial(self._removePlot, wdg))
+            self._plots[plot] = wdg
+
+        return self._plots[plot]
+
+
+class PlotWidget(QFrame, Ui_PlotWidget, EventListener):
+    titleChanged = pyqtSignal()
+    iconChanged = pyqtSignal()
+    characterChanged = pyqtSignal()
+    removalRequested = pyqtSignal()
+
+    def __init__(self, novel: Novel, plot: Plot, parent=None):
+        super(PlotWidget, self).__init__(parent)
+        self.setupUi(self)
+        self.novel = novel
+        self.plot: Plot = plot
+
+        incr_font(self.lineName, 6)
+        bold(self.lineName)
+        self.lineName.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        self.lineName.setText(self.plot.text)
+        self.lineName.textChanged.connect(self._nameEdited)
+
+        self.btnPincipleEditor.setIcon(IconRegistry.plus_edit_icon('grey'))
+        transparent(self.btnPincipleEditor)
+        retain_when_hidden(self.btnPincipleEditor)
+        decr_icon(self.btnPincipleEditor)
+
+        self._principleSelectorMenu = PlotPrincipleSelectorMenu(self.plot, self.btnPincipleEditor)
+        self._principleSelectorMenu.principleToggled.connect(self._principleToggled)
+        self._principleSelectorMenu.progressionToggled.connect(self._progressionToggled)
+        self._principleSelectorMenu.genresSelected.connect(self._genresSelected)
+        self.btnPincipleEditor.installEventFilter(ButtonPressResizeEventFilter(self.btnPincipleEditor))
+        self.btnPincipleEditor.installEventFilter(OpacityEventFilter(self.btnPincipleEditor, leaveOpacity=0.7))
+        self._principles: Dict[PlotPrincipleType, PlotPrincipleEditor] = {}
+
+        self._initFrameColor()
+        self.btnPrinciples.setIcon(IconRegistry.from_name('mdi6.note-text-outline', 'grey'))
+        incr_icon(self.btnPrinciples, 2)
+        incr_font(self.btnPrinciples, 2)
+        self.btnPrinciples.installEventFilter(ButtonPressResizeEventFilter(self.btnPrinciples))
+        self.btnPrinciples.installEventFilter(OpacityEventFilter(self.btnPrinciples, leaveOpacity=0.7))
+        self.btnPrinciples.clicked.connect(lambda: self._principleSelectorMenu.exec())
+
+        flow(self.wdgPrinciples, spacing=6)
+        margins(self.wdgPrinciples, left=30)
+        for principle in self.plot.principles:
+            self._initPrincipleEditor(principle)
+
+        self.btnProgression.setIcon(IconRegistry.rising_action_icon('grey'))
+        if self.plot.plot_type == PlotType.Internal:
+            self.btnProgression.setText('Transformation')
+        elif self.plot.plot_type == PlotType.Relation:
+            self.btnProgression.setText('Evolution')
+
+        translucent(self.btnProgression, 0.7)
+        incr_icon(self.btnProgression)
+        incr_font(self.btnProgression)
+
+        self.btnValues.setText('' if self.plot.values else 'Values')
+        self.btnValues.setIcon(IconRegistry.from_name('fa5s.chevron-circle-down', 'grey'))
+        self.btnValues.installEventFilter(OpacityEventFilter(self.btnValues, 0.9, 0.7))
+        self.btnValues.clicked.connect(self._newValue)
+        hbox(self.wdgValues)
+        self._btnAddValue = SecondaryActionPushButton(self)
+        self._btnAddValue.setFocusPolicy(Qt.FocusPolicy.ClickFocus)
+        decr_font(self._btnAddValue)
+        self._btnAddValue.setIconSize(QSize(14, 14))
+        retain_when_hidden(self._btnAddValue)
+        self._btnAddValue.setIcon(IconRegistry.plus_icon('grey'))
+        for value in self.plot.values:
+            self._addValue(value)
+
+        self.btnRelationArrow.setHidden(True)
+        self._characterRelationSelector: Optional[CharacterAvatar] = None
+        if self.plot.plot_type == PlotType.Global:
+            pass
+        elif self.plot.plot_type == PlotType.Relation:
+            self.btnRelationArrow.setVisible(True)
+            self.btnRelationArrow.setIcon(IconRegistry.from_name('ph.arrows-counter-clockwise-fill'))
+            self.btnPlotIcon.setIconSize(QSize(32, 32))
+
+            self._characterSelector = CharacterAvatar(self, 60, 100, 64, 8)
+            self._characterRelationSelector = CharacterAvatar(self, 60, 100, 64, 8)
+            self._characterSelector.setToolTip('Associate a character to this relationship plot')
+            self._characterRelationSelector.setToolTip('Associate a character to this relationship plot')
+            sourceMenu = CharacterSelectorMenu(self.novel, self._characterSelector.btnAvatar)
+            sourceMenu.selected.connect(self._characterSelected)
+            targetMenu = CharacterSelectorMenu(self.novel, self._characterRelationSelector.btnAvatar)
+            targetMenu.selected.connect(self._relationCharacterSelected)
+            self._characterSelector.setFixedSize(90, 90)
+            self._characterRelationSelector.setFixedSize(90, 90)
+            self.wdgHeader.layout().insertWidget(0, self._characterSelector,
+                                                 alignment=Qt.AlignmentFlag.AlignCenter)
+            self.wdgHeader.layout().insertWidget(0, spacer())
+            self.wdgHeader.layout().addWidget(self._characterRelationSelector, alignment=Qt.AlignmentFlag.AlignCenter)
+            self.wdgHeader.layout().addWidget(spacer())
+
+            character = self.plot.character(novel)
+            if character is not None:
+                self._characterSelector.setCharacter(character)
+            character = self.plot.relation_character(novel)
+            if character is not None:
+                self._characterRelationSelector.setCharacter(character)
+        else:
+            self._characterSelector = CharacterAvatar(self, 88, 120, 92, 8)
+            menu = CharacterSelectorMenu(self.novel, self._characterSelector.btnAvatar)
+            menu.selected.connect(self._characterSelected)
+            self._characterSelector.setToolTip('Link character to this storyline')
+            self._characterSelector.setGeometry(20, 20, 115, 115)
+            character = self.plot.character(novel)
+            if character is not None:
+                self._characterSelector.setCharacter(character)
+
+        self.wdgValues.layout().addWidget(self._btnAddValue)
+        self.wdgValues.layout().addWidget(spacer())
+        self._btnAddValue.clicked.connect(self._newValue)
+
+        self.installEventFilter(VisibilityToggleEventFilter(target=self.btnSettings, parent=self))
+        self.installEventFilter(VisibilityToggleEventFilter(target=self._btnAddValue, parent=self))
+        self.installEventFilter(VisibilityToggleEventFilter(target=self.btnPincipleEditor, parent=self))
+
+        self.btnPlotIcon.installEventFilter(OpacityEventFilter(self.btnPlotIcon, enterOpacity=0.7, leaveOpacity=1.0))
+        self._updateIcon()
+
+        self._timeline = PlotEventsTimeline(self.novel, self.plot.plot_type)
+        self.wdgEventsParent.layout().addWidget(self._timeline)
+        self._timeline.setStructure(self.plot.progression)
+        self._timeline.timelineChanged.connect(self._timelineChanged)
+
+        self.wdgProgression.setVisible(self.plot.has_progression)
+
+        iconMenu = MenuWidget(self.btnPlotIcon)
+
+        colorPicker = ColorPicker(self)
+        colorPicker.setFixedSize(300, 150)
+        colorPicker.colorPicked.connect(self._colorChanged)
+        colorMenu = MenuWidget()
+        colorMenu.setTitle('Color')
+        colorMenu.setIcon(IconRegistry.from_name('fa5s.palette'))
+        colorMenu.addWidget(colorPicker)
+
+        iconMenu.addMenu(colorMenu)
+        iconMenu.addSeparator()
+        iconMenu.addAction(
+            action('Change icon', icon=IconRegistry.icons_icon(), slot=self._changeIcon, parent=iconMenu))
+
+        contextMenu = MenuWidget(self.btnSettings)
+        progress_action = action('Track general progress',
+                                 slot=self._trackGeneralProgressChanged,
+                                 checkable=True,
+                                 tooltip='Enable tracking a general progression value besides the custom plot values',
+                                 parent=contextMenu)
+        progress_action.setChecked(self.plot.default_value_enabled)
+        contextMenu.addAction(progress_action)
+        contextMenu.addSeparator()
+        contextMenu.addAction(action('Remove plot', IconRegistry.trash_can_icon(), self.removalRequested.emit))
+
+        self.repo = RepositoryPersistenceManager.instance()
+
+        dispatcher = event_dispatchers.instance(self.novel)
+        dispatcher.register(self, CharacterChangedEvent, CharacterDeletedEvent)
+
+    @overrides
+    def event_received(self, event: Event):
+        if isinstance(event, CharacterDeletedEvent):
+            if self._characterSelector.character() and self._characterSelector.character().id == event.character.id:
+                self._characterSelector.reset()
+            if self._characterRelationSelector and self._characterRelationSelector.character() and self._characterRelationSelector.character().id == event.character.id:
+                self._characterRelationSelector.reset()
+
+    def _updateIcon(self):
+        if self.plot.icon:
+            self.btnPlotIcon.setIcon(IconRegistry.from_name(self.plot.icon, self.plot.icon_color))
+
+    def _nameEdited(self, name: str):
+        self.plot.text = name
+        self._save()
+        self.titleChanged.emit()
+
+    def _characterSelected(self, character: Character):
+        self._characterSelector.setCharacter(character)
+        self.plot.set_character(character)
+        self._save()
+        self.characterChanged.emit()
+
+    def _relationCharacterSelected(self, character: Character):
+        self._characterRelationSelector.setCharacter(character)
+        self.plot.set_relation_character(character)
+        self._save()
+
+    def _changeIcon(self):
+        result = IconSelectorDialog(self).display(QColor(self.plot.icon_color))
+        if result:
+            self.plot.icon = result[0]
+            self._colorChanged(result[1])
+
+    def _colorChanged(self, color: QColor):
+        self.plot.icon_color = color.name()
+        self._updateIcon()
+        self._initFrameColor()
+        self._save()
+        self.iconChanged.emit()
+
+    def _trackGeneralProgressChanged(self, toggled: bool):
+        self.plot.default_value_enabled = toggled
+        self._save()
+
+    def _principleToggled(self, principleType: PlotPrincipleType, toggled: bool):
+        if toggled:
+            principle = PlotPrinciple(principleType)
+            self._initPrincipleEditor(principle)
+            self.plot.principles.append(principle)
+        else:
+            principle = next((_principle for _principle in self.plot.principles if _principle.type == principleType),
+                             None)
+            if principle:
+                self.plot.principles.remove(principle)
+                wdg = self._principles.pop(principle.type)
+                fade_out_and_gc(self.wdgPrinciples, wdg)
+
+        self._btnAddValue.setVisible(True)
+        self.btnSettings.setVisible(True)
+        self.btnPincipleEditor.setVisible(True)
+
+        self._save()
+
+    def _progressionToggled(self, toggled: bool):
+        self.plot.has_progression = toggled
+        self.wdgProgression.setVisible(self.plot.has_progression)
+        self._save()
+
+    def _genresSelected(self):
+        object = PrincipleSelectorObject()
+        object.principleToggled.connect(self._principleToggled)
+        GenrePrincipleSelectorDialog.popup(self.plot, object)
+
+    def _initPrincipleEditor(self, principle: PlotPrinciple):
+        editor = PlotPrincipleEditor(principle, self.plot.plot_type)
+        editor.principleEdited.connect(self._save)
+        # self.wdgPrinciples.layout().insertWidget(principle_type_index[principle.type], editor)
+        self.wdgPrinciples.layout().insertWidget(principle.type.value, editor)
+        self._principles[principle.type] = editor
+
+        return editor
+
+    def _save(self):
+        self.repo.update_novel(self.novel)
+
+    def _timelineChanged(self):
+        self._save()
+
+    def _initFrameColor(self):
+        self.setStyleSheet(f'''
+            #PlotWidget {{
+                background-color: {RELAXED_WHITE_COLOR};
+                border: 2px solid {self.plot.icon_color};
+            }}
+            #scrollAreaWidgetContents {{
+                background-color: {RELAXED_WHITE_COLOR};
+            }}
+            #frame {{
+                border: 1px solid {self.plot.icon_color};
+            }}
+        ''')
+
+    def _newValue(self):
+        value = PlotValueEditorDialog().display()
+        if value:
+            self.plot.values.append(value)
+            self.wdgValues.layout().removeWidget(self._btnAddValue)
+            self._addValue(value)
+            self.wdgValues.layout().addWidget(self._btnAddValue)
+
+            self._save()
+
+    def _addValue(self, value: PlotValue):
+        label = PlotValueLabel(value, parent=self.wdgValues, simplified=True)
+        sp(label).h_max()
+        label.installEventFilter(OpacityEventFilter(label, leaveOpacity=0.7))
+        pointy(label)
+        insert_before_the_end(self.wdgValues, label)
+        label.removalRequested.connect(partial(self._removeValue, label))
+        label.clicked.connect(partial(self._plotValueClicked, label))
+
+        self.btnValues.setText('')
+
+    def _removeValue(self, label: PlotValueLabel):
+        if app_env.test_env():
+            self.__destroyValue(label)
+        else:
+            anim = qtanim.fade_out(label, duration=150, hide_if_finished=False)
+            anim.finished.connect(partial(self.__destroyValue, label))
+
+    def _editValue(self, label: PlotValueLabel):
+        value = PlotValueEditorDialog().display(label.value)
+        if value:
+            label.value.text = value.text
+            label.value.negative = value.negative
+            label.value.icon = value.icon
+            label.value.icon_color = value.icon_color
+            label.refresh()
+            self._save()
+
+    def _plotValueClicked(self, label: PlotValueLabel):
+        menu = MenuWidget()
+        menu.addAction(action('Edit', IconRegistry.edit_icon(), partial(self._editValue, label)))
+        menu.addSeparator()
+        menu.addAction(action('Remove', IconRegistry.trash_can_icon(), label.removalRequested.emit))
+        menu.exec(QCursor.pos())
+
+    def __destroyValue(self, label: PlotValueLabel):
+        self.plot.values.remove(label.value)
+        self._save()
+        fade_out_and_gc(self.wdgValues, label)
+        self.btnValues.setText('' if self.plot.values else 'Values')
+
+
+class PlotEditor(QWidget, Ui_PlotEditor):
+    def __init__(self, novel: Novel, parent=None):
+        super(PlotEditor, self).__init__(parent)
+        self.setupUi(self)
+        self.novel = novel
+
+        self._wdgList = PlotTreeView(self.novel)
+        self.wdgPlotListParent.layout().addWidget(self._wdgList)
+        self._wdgList.plotSelected.connect(self._plotSelected)
+        self._wdgList.plotRemoved.connect(self._plotRemoved)
+        self.stack.setCurrentWidget(self.pageDisplay)
+
+        self._wdgImpactMatrix = StorylinesImpactMatrix(self.novel)
+        self.scrollMatrix.layout().addWidget(self._wdgImpactMatrix)
+
+        self.splitter.setSizes([150, 550])
+
+        italic(self.btnAdd)
+        self.btnAdd.setIcon(IconRegistry.plus_icon('white'))
+        self.btnImpactMatrix.setIcon(IconRegistry.from_name('mdi6.camera-metering-matrix'))
+        self.btnImpactMatrix.clicked.connect(self._displayImpactMatrix)
+
+        menu = MenuWidget(self.btnAdd)
+        menu.addAction(action('Main plot', IconRegistry.storylines_icon(), lambda: self.newPlot(PlotType.Main)))
+        menu.addAction(
+            action('Character arc', IconRegistry.conflict_self_icon(), lambda: self.newPlot(PlotType.Internal)))
+        menu.addAction(action('Subplot', IconRegistry.subplot_icon(), lambda: self.newPlot(PlotType.Subplot)))
+
+        submenu = MenuWidget()
+        submenu.setTitle('Other')
+        submenu.addAction(action('Relationship plot', IconRegistry.from_name('fa5s.people-arrows'),
+                                 slot=lambda: self.newPlot(PlotType.Relation)))
+        submenu.addAction(action('Global storyline', IconRegistry.from_name('fa5s.globe'),
+                                 slot=lambda: self.newPlot(PlotType.Global)))
+        menu.addSeparator()
+        menu.addMenu(submenu)
+
+        self.repo = RepositoryPersistenceManager.instance()
+
+        if self.novel.plots:
+            self._wdgList.selectPlot(self.novel.plots[0])
+
+    def widgetList(self) -> PlotTreeView:
+        return self._wdgList
+
+    def newPlot(self, plot_type: PlotType):
+        if plot_type == PlotType.Internal:
+            name = 'Character arc'
+            icon = 'mdi.mirror'
+        elif plot_type == PlotType.Subplot:
+            name = 'Subplot'
+            icon = 'mdi.source-branch'
+        elif plot_type == PlotType.Relation:
+            name = 'Relationship'
+            icon = 'fa5s.people-arrows'
+        elif plot_type == PlotType.Global:
+            name = 'Global storyline'
+            icon = 'fa5s.globe'
+        else:
+            name = 'Main plot'
+            icon = 'fa5s.theater-masks'
+        plot = Plot(name, plot_type=plot_type, icon=icon,
+                    progression=[PlotProgressionItem(type=PlotProgressionItemType.BEGINNING),
+                                 PlotProgressionItem(type=PlotProgressionItemType.MIDDLE),
+                                 PlotProgressionItem(type=PlotProgressionItemType.ENDING)])
+        self.novel.plots.append(plot)
+
+        plot_colors = list(STORY_LINE_COLOR_CODES[plot_type.value])
+        for plot in self.novel.plots:
+            if plot.plot_type == plot_type and plot.icon_color in plot_colors:
+                plot_colors.remove(plot.icon_color)
+        if plot_colors:
+            plot.icon_color = plot_colors[0]
+        else:
+            plot_colors = STORY_LINE_COLOR_CODES[plot_type.value]
+            number_of_plots = len([x for x in self.novel.plots if x.plot_type == plot_type])
+            plot.icon_color = plot_colors[(number_of_plots - 1) % len(plot_colors)]
+
+        self._wdgList.addPlot(plot)
+        self.repo.update_novel(self.novel)
+        self._wdgList.selectPlot(plot)
+        self._wdgImpactMatrix.refresh()
+
+        emit_event(self.novel, StorylineCreatedEvent(self))
+
+    def _plotSelected(self, plot: Plot) -> PlotWidget:
+        self.btnImpactMatrix.setChecked(False)
+        self.stack.setCurrentWidget(self.pageDisplay)
+
+        widget = PlotWidget(self.novel, plot, self.pageDisplay)
+        widget.removalRequested.connect(partial(self._remove, widget))
+        widget.titleChanged.connect(partial(self._wdgList.refreshPlot, widget.plot))
+        widget.iconChanged.connect(partial(self._wdgList.refreshPlot, widget.plot))
+        widget.characterChanged.connect(self._wdgList.refreshCharacters)
+
+        clear_layout(self.pageDisplay)
+        self.pageDisplay.layout().addWidget(widget)
+
+        return widget
+
+    def _remove(self, wdg: PlotWidget):
+        self._wdgList.removePlot(wdg.plot)
+
+    def _plotRemoved(self, plot: Plot):
+        if self.pageDisplay.layout().count():
+            item = self.pageDisplay.layout().itemAt(0)
+            if item.widget() and isinstance(item.widget(), PlotWidget):
+                if item.widget().plot == plot:
+                    clear_layout(self.pageDisplay)
+        delete_plot(self.novel, plot)
+
+        self._wdgImpactMatrix.refresh()
+        emit_event(self.novel, StorylineRemovedEvent(self, plot))
+
+    # def _remove(self, widget: PlotWidget):
+    #     if ask_confirmation(f'Are you sure you want to delete the plot {widget.plot.text}?'):
+    #         if app_env.test_env():
+    #             self.__destroy(widget)
+    #         else:
+    #             anim = qtanim.fade_out(widget, duration=150)
+    #             anim.finished.connect(partial(self.__destroy, widget))
+    #
+    # def __destroy(self, widget: PlotWidget):
+    #     delete_plot(self.novel, widget.plot)
+    #     self.scrollAreaWidgetContents.layout().removeWidget(widget.parent())
+
+    def _displayImpactMatrix(self, checked: bool):
+        self._wdgList.clearSelection()
+        if checked:
+            self.stack.setCurrentWidget(self.pageMatrix)
+        else:
+            self.stack.setCurrentWidget(self.pageDisplay)
